requirements.txt
Django>=5.1
djangorestframework>=3.15
django-cors-headers>=4.6
openpyxl>=3.1
Pillow>=12.0.0
psycopg2-binary>=2.9.11
qrcode